+++ src/main/java/subway/line/presentation/request/LineUpdateRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.presentation.request;
 
 import lombok.Getter;
